+++ drake/examples/spring_mass/spring_mass_system.cc
@@ -2,9 +2,11 @@
 
 namespace drake {
 
+using systems::ContextBase;
 using systems::BasicVector;
 using systems::Context;
 using systems::ContinuousState;
+using systems::LeafSystemOutput;
 using systems::OutputPort;
 using systems::StateVector;
 using systems::SystemOutput;
@@ -46,7 +48,8 @@ SpringMassStateVector* SpringMassStateVector::DoClone() const {
 }
 
 SpringMassOutputVector::SpringMassOutputVector()
-    : BasicVector<double>(kStateSize-1) {}  // don't output conservative energy
+    : BasicVector<double>(kStateSize - 1) {
+}  // don't output conservative energy
 
 double SpringMassOutputVector::get_position() const { return get_value()[0]; }
 double SpringMassOutputVector::get_velocity() const { return get_value()[1]; }
@@ -74,25 +77,21 @@ SpringMassSystem::SpringMassSystem(const std::string& name,
       system_is_forced_(system_is_forced) {}
 
 double SpringMassSystem::EvalSpringForce(const MyContext& context) const {
-  const double k = spring_constant_N_per_m_,
-               x = get_position(context),
+  const double k = spring_constant_N_per_m_, x = get_position(context),
                x0 = 0.,  // TODO(david-german-tri) should be a parameter.
-               stretch = x - x0, f = -k * stretch;
+      stretch = x - x0, f = -k * stretch;
   return f;
 }
 
 double SpringMassSystem::EvalPotentialEnergy(const MyContext& context) const {
-  const double k = spring_constant_N_per_m_,
-               x = get_position(context),
+  const double k = spring_constant_N_per_m_, x = get_position(context),
                x0 = 0.,  // TODO(david-german-tri) should be a parameter.
-               stretch = x - x0, pe = k * stretch * stretch / 2;
+      stretch = x - x0, pe = k * stretch * stretch / 2;
   return pe;
 }
 
 double SpringMassSystem::EvalKineticEnergy(const MyContext& context) const {
-  const double m = mass_kg_,
-               v = get_velocity(context),
-               ke = m * v * v / 2;
+  const double m = mass_kg_, v = get_velocity(context), ke = m * v * v / 2;
   return ke;
 }
 
@@ -109,26 +108,28 @@ double SpringMassSystem::EvalNonConservativePower(const MyContext&) const {
 std::string SpringMassSystem::get_name() const { return name_; }
 
 // Reserve a context with no input, and a SpringMassStateVector state.
-std::unique_ptr<Context<double>> SpringMassSystem::CreateDefaultContext()
-    const {
+std::unique_ptr<ContextBase<double>>
+SpringMassSystem::CreateDefaultContext() const {
   std::unique_ptr<Context<double>> context(new Context<double>);
   std::unique_ptr<SpringMassStateVector> state(new SpringMassStateVector(0, 0));
   context->get_mutable_state()->continuous_state.reset(
       new ContinuousState<double>(std::move(state), 1 /* size of q */,
                                   1 /* size of v */, 1 /* size of z */));
   if (system_is_forced_) context->SetNumInputPorts(1);
-  return context;
+  return std::unique_ptr<ContextBase<double>>(context.release());
 }
 
-std::unique_ptr<SystemOutput<double>> SpringMassSystem::AllocateOutput() const {
-  std::unique_ptr<SystemOutput<double>> output(new SystemOutput<double>);
+std::unique_ptr<SystemOutput<double>> SpringMassSystem::AllocateOutput(
+    const ContextBase<double>& context) const {
+  std::unique_ptr<LeafSystemOutput<double>> output(
+      new LeafSystemOutput<double>);
   {
     std::unique_ptr<VectorInterface<double>> data(new SpringMassOutputVector());
     std::unique_ptr<OutputPort<double>> port(
         new OutputPort<double>(std::move(data)));
-    output->ports.push_back(std::move(port));
+    output->get_mutable_ports()->push_back(std::move(port));
   }
-  return output;
+  return std::unique_ptr<SystemOutput<double>>(output.release());
 }
 
 std::unique_ptr<ContinuousState<double>>
@@ -141,7 +142,7 @@ SpringMassSystem::AllocateTimeDerivatives() const {
 }
 
 // Assign the state to the output.
-void SpringMassSystem::EvalOutput(const Context<double>& context,
+void SpringMassSystem::EvalOutput(const ContextBase<double>& context,
                                   SystemOutput<double>* output) const {
   // TODO(david-german-tri): Cache the output of this function.
   const SpringMassStateVector& state = get_state(context);
@@ -152,7 +153,7 @@ void SpringMassSystem::EvalOutput(const Context<double>& context,
 
 // Compute the actual physics.
 void SpringMassSystem::EvalTimeDerivatives(
-    const Context<double>& context,
+    const ContextBase<double>& context,
     ContinuousState<double>* derivatives) const {
   // TODO(david-german-tri): Cache the output of this function.
   const SpringMassStateVector& state = get_state(context);