+++ drake/solvers/MosekSolver.cpp
@@ -1,9 +1,9 @@
 // Copyright 2016, Alex Dunyak
-// A wrapper file for MosekLP and mosekQP that handles constraint and
+// A wrapper file for MosekWrapper and mosekQP that handles constraint and
 // objective marshalling
 
 #include "drake/solvers/MosekSolver.h"
-#include "drake/solvers/MosekLP.h"
+#include "drake/solvers/MosekWrapper.h"
 
 #include <Eigen/Core>
 
@@ -19,8 +19,10 @@ bool MosekSolver::available() const { return true; }
 SolutionResult MosekSolver::Solve(OptimizationProblem& prog) const {
   if (!prog.GetSolverOptionsStr("Mosek").empty()) {
     if (prog.GetSolverOptionsStr("Mosek").at("problemtype").find("linear")
-        != std::string::npos) {
-      return MosekLP::Solve(prog);
+            != std::string::npos ||
+        prog.GetSolverOptionsStr("Mosek").at("problemtype").find("quadratic")
+            != std::string::npos) {
+      return MosekWrapper::Solve(prog);
     }
   }  // TODO(alexdunyak): add more mosek solution types.
   return kUnknownError;