+++ drake/systems/plants/parser_urdf.cc
@@ -387,7 +387,6 @@ void parseVisual(RigidBody* body, XMLElement* node, RigidBodyTree* tree,
     {
       XMLElement* color_node = material_node->FirstChildElement("color");
       if (color_node) {
-        Vector4d rgba;
         if (!parseVectorAttribute(color_node, "rgba", rgba)) {
           throw runtime_error(
               "ERROR: Failed to parse color of material for "
@@ -489,7 +488,7 @@ void parseCollision(RigidBody* body, XMLElement* node, RigidBodyTree* tree,
     throw runtime_error("ERROR: Link " + body->get_name() +
                         " has a collision element without geometry");
 
-  RigidBody::CollisionElement element(T_element_to_link, body);
+  RigidBodyCollisionElement element(T_element_to_link, body);
   // By default all collision elements added to the world from an URDF file are
   // flagged as static.
   // We would also like to flag as static bodies connected to the world with a
@@ -694,7 +693,7 @@ void parseJoint(RigidBodyTree* tree, XMLElement* node) {
 
   unique_ptr<DrakeJoint> joint_unique_ptr(joint);
   tree->bodies[child_index]->setJoint(move(joint_unique_ptr));
-  tree->bodies[child_index]->parent = tree->bodies[parent_index].get();
+  tree->bodies[child_index]->set_parent(tree->bodies[parent_index].get());
 }
 
 // Searches through the URDF document looking for the effort limits of a