+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Unreleased: changes on master, not yet released
 ### Changed
 
  - [#2923][] Updated member variables of `RigidBodyLoop` and `RigidBodActuator` to conform to style guide.
+ - [#2904][] Made `RigidBody::parent` private. Re-named it to be `RigidBody::parent_`. Added necessary accessors.
  - [#2903][] Made `RigidBody::robotnum` private. Re-named it to be `RigidBody::model_id_`.
  - [#2902][] Made `RigidBody::model_name_` private. Re-named `RigidBody::model_name()` to be `RigidBody::get_model_name()`. Added `RigidBody::set_model_name()`.
  - [#2900][] Made `RigidBody::name_` private. Re-named `RigidBody::name()` to be `RigidBody::get_name()`. Added `RigidBody::set_name()`.
@@ -38,6 +39,7 @@ Unreleased: changes on master, not yet released
  - [#2018][] Fix capitalization of `Constraint` and `OptimizationProblem` APIs to match style guide.
  - [#2415][] Require CMake 3.5 or higher to configure and build.
  - [#2779][] Move some rotation functions from drakeGeometryUtil to drake/math.
+ - [#2963][] Rename RigidBody::CollisionElement to RigidBodyCollisionElement.
 
 [//]: # "Lost functionality or APIs."
 ### Removed / Deprecated
@@ -89,4 +91,6 @@ Changes in version v0.9.11 and before are not provided.
 [#2900]: https://github.com/RobotLocomotion/drake/issues/2900
 [#2902]: https://github.com/RobotLocomotion/drake/issues/2902
 [#2903]: https://github.com/RobotLocomotion/drake/issues/2903
+[#2904]: https://github.com/RobotLocomotion/drake/issues/2904
 [#2923]: https://github.com/RobotLocomotion/drake/issues/2923
+[#2963]: https://github.com/RobotLocomotion/drake/issues/2963